TreeMap Flush to Disk on limit reached.
The data is inserted into treemap and a limit is set to tree map. When the TreeMap limit is reached. The program will create a file and flush all the data in TreeMap to file and then it will clear the treemap.
When the searching for particular string begins. program first searches in TreeMap if not found searches in file and if found returns the string. Else displays a message. String not found.